reorganize prefab related files
