+++ cocos/core/utils/prefab/index.ts
@@ -1,0 +1,2 @@
+export * from './prefab-info';
+export * from './utils';
